little changes

+++ app/src/main/java/com/example/mobilelab/model/server/form/TaskForm.kt
@@ -4,7 +4,7 @@ data class TaskForm(
     val title: String,
     val description: String,
     val done: Int,
-    val deadline: Int,
+    val deadline: Long,
     val category_id: Int,
     val priority_id: Int
 )
